Extracting navtabs as separated React Components
