+++ src/js/constants/tabs.js
@@ -1,6 +1,20 @@
+import React from "react/addons";
+import DeploymentsListComponent
+  from "../components/DeploymentsListComponent";
+import AppsAndGroupsListComponent
+  from "../components/AppsAndGroupsListComponent";
+
 var tabs = [
-  {id: "/apps", text: "Applications"},
-  {id: "/deployments", text: "Deployments"}
+  {
+    id: "/apps",
+    text: "Applications",
+    component: AppsAndGroupsListComponent
+  },
+  {
+    id: "/deployments",
+    text: "Deployments",
+    component: DeploymentsListComponent
+  }
 ];
 
 export default tabs;
